Don't register empty anonymous concepts by default and auto register anonymous concepts when referenced. Drop equivalent annotation as it can be done by an OWL inferencer.
git-svn-id: http://repo.aduna-software.org/svn/org.openrdf/alibaba/trunk@8892 3fbd4d3e-0f96-47c4-bab7-89622e02a19e

+++ object-repository/src/main/java/org/openrdf/repository/object/ObjectRepository.java
@@ -185,7 +185,7 @@ public class ObjectRepository extends ContextAwareRepository {
 	@Override
 	public void shutDown() throws StoreException {
 		super.shutDown();
-		if (getDataDir() == null) {
+		if (getDataDir() == null && dataDir != null) {
 			try {
 				FileUtil.deleteDir(dataDir);
 			} catch (IOException e) {
